+++ use-cases/roles/delete-role.js
@@ -1,5 +1,4 @@
-const fromEntities = require("../../entity");
-
+const fromEntities = require('../../entity');
 
 exports.Delete = ({
   CreateError,
@@ -10,7 +9,7 @@ exports.Delete = ({
   request,
   db,
   ac,
-  accessManager
+  accessManager,
 }) => {
   return Object.freeze({
     execute: async () => {
@@ -20,19 +19,18 @@ exports.Delete = ({
         const userUID = request.locals.uid;
         const role = request.locals.role;
         const id = request.queryParams.id;
-        const acesssRes = await accessManager({
-          translate,
-          logger,
-          CreateError,
-          lang,
-          role,
-          db,
-          useCase: 'roles:edit',
-        })
-        if(!acesssRes)
-        {
-          throw new CreateError(translate(lang, "forbidden"), 403);
-        }
+        // const acesssRes = await accessManager({
+        //   translate,
+        //   logger,
+        //   CreateError,
+        //   lang,
+        //   role,
+        //   db,
+        //   useCase: 'roles:edit',
+        // });
+        // if (!acesssRes) {
+        //   throw new CreateError(translate(lang, 'forbidden'), 403);
+        // }
 
         // let permission = ac.can(role).createOwn("mood");
 
@@ -44,17 +42,32 @@ exports.Delete = ({
         //   throw new CreateError(translate(lang, "forbidden"), 403);
         // }
 
-const RoleFunction =db.methods.Role({
-  translate,
-  logger,
-  CreateError,
-  lang,
-})
+        const RoleFunction = db.methods.Role({
+          translate,
+          logger,
+          CreateError,
+          lang,
+        });
 
-const res = await RoleFunction.deleteById({id})
+        const UserFunction = db.methods.User({
+          translate,
+          logger,
+          CreateError,
+          lang,
+        });
+
+        const users = await UserFunction.findByParams({
+          role: id,
+        });
+
+        if (users.data.length > 0) {
+          throw new CreateError(translate(lang, 'forbidden_request'), 403);
+        }
+
+        const res = await RoleFunction.deleteById({ id });
         return {
-          msg: translate(lang, "created_mood"),
-          data: { res},
+          msg: translate(lang, 'created_mood'),
+          data: { res },
         };
       } catch (error) {
         if (error instanceof CreateError) {
